Move Suggest global and provider config to config.rs
