+++ components/suggest/src/lib.rs
@@ -4,6 +4,7 @@
  */
 
 use remote_settings::RemoteSettingsConfig;
+mod config;
 mod db;
 mod error;
 mod keyword;
@@ -15,12 +16,10 @@ mod store;
 mod suggestion;
 mod yelp;
 
+pub use config::{SuggestGlobalConfig, SuggestProviderConfig};
 pub use error::SuggestApiError;
 pub use provider::SuggestionProvider;
-pub use store::{
-    SuggestGlobalConfig, SuggestIngestionConstraints, SuggestProviderConfig, SuggestStore,
-    SuggestStoreBuilder,
-};
+pub use store::{SuggestIngestionConstraints, SuggestStore, SuggestStoreBuilder};
 pub use suggestion::{raw_suggestion_url_matches, Suggestion};
 
 pub(crate) type Result<T> = std::result::Result<T, error::Error>;
